pull: resolve the image config and set it on stored image
Previously `pull` would not properly retain the image configuration
which contains the entrypoint, working directory, user, and more. This
resulted in broken images if they were pulled then subsequently pushed.

This commit fetches the image config, which must be done separately, and
then ensures that the configuration is set on the exported image into
the local img state store.